contato_django/requirements.txt
Django>=4.2,<5.0
pillow
djangorestframework
